requirements.txt
pymupdf
python-docx
pytesseract
pillow
jinja2
pdfkit
requests